backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.2",
    "simpy>=4.1.1",
    "pydantic>=2.5.0",
    " jinja2>=3.1.0",
    "google-generativeai>=0.8.0",
    "torch>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
backend = "backend.main:main"

[tool.uv]
package = true

[tool.uv.sources]
backend = { path = "src" }
